requirements.txt
openai==1.12.0
ffmpeg-python==0.2.0
python-dotenv==1.0.0
pydub==0.25.
assemblyai>=0.19.0
